cleanup some code fix sinepath
made generate projectile its own function to make code more readable.
changed sine function to cosine so that it would start the projectile in the middle of a wave instead of at the bottom. this way it the average position of the projectile on the same axis as the origin (player, enemy, etc)